Add hitbox delay and animation

+++ objects/oHitbox/Step_0.gml
@@ -3,18 +3,23 @@ move.hsp = approach(move.hsp, 0, move.fric);
 move.vsp = approach(move.vsp, 0, move.fric);
 move.dir = point_direction(0, 0, move.hsp, move.vsp);
 
-if (destroyOnStop && move.hsp == 0 && move.vsp == 0) destroySelf();
+if (atk.destroyOnStop && move.hsp == 0 && move.vsp == 0) destroySelf();
 
 x += move.hsp * delta;
 y += move.vsp * delta;
 
-if (piercing) {
-	getTouchingObjects(atk.damagedEnemies, oEnemyBase, dealDamage);
-} else {
-	var enemy = instance_place(x, y, oEnemyBase);
-	if (enemy != noone) {
-		dealDamage(enemy);
-		destroySelf();
+//Hitbox only becomes active when delay is over
+atk.delay = approach(atk.delay, 0, 1);
+
+if (atk.delay == 0) {
+	if (atk.piercing) {
+		getTouchingObjects(atk.damagedEnemies, oEnemyBase, dealDamage);
+	} else {
+		var enemy = instance_place(x, y, oEnemyBase);
+		if (enemy != noone) {
+			dealDamage(enemy);
+			destroySelf();
+		}
 	}
 }
 
@@ -23,3 +28,7 @@ depthSorting();
 //See if boi needs to be destroyed
 atk.dur = approach(atk.dur, 0, 1);
 if (atk.dur <= 0) destroySelf();
+
+//Animation
+visuals.frm = approach(visuals.frm, image_number - 1, visuals.animSpd);
+image_index = floor(visuals.frm);
